Handle POWER kernel return values
The current code has a very x86 centric check to see if the return
value of the vDSO functions returned a negative errno which should be
translated into a -1 return value and a positive errno.

POWER indicates the success or failure of a syscall by using a
condition register bit to indicate failure and then returning a
positive errno directly. Unfortunately reading the condition register
in C is almost impossible.

Introduce a macro to wrap calls to the vdso which can return both the
condition register and the return value of the vDSO call back into C
for POWER.

Leave current behaviour unchanged for other architectures.